requirements.txt
streamlit
requests
pyyaml
openai
certifi
streamlit_chat
python-dotenv
azure-ai-vision-imageanalysis